requirement.txt
fastapi
aptos-sdk
"uvicorn[standard]"
apscheduler
numpy
python-dotenv